requirements.txt
google-generativeai  
python-telegram-bot[socks]  
psycopg2-binary
python-dotenv  
langchain  
langchain-google-genai  
google-api-python-client
python-dateutil